Move the main Koto interface into its own file
